Initial commit for hyperopt implementation
