+++ src/flownet/_command_line.py
@@ -4,7 +4,8 @@ import pathlib
 import subprocess
 
 from .config_parser import parse_config, parse_pred_config
-from .ahm import run_flownet_history_matching
+from .ahm import run_esmda
+from .ahm import run_hyperopt
 from .prediction import run_flownet_prediction
 
 
@@ -44,9 +45,9 @@ def create_webviz(output_folder: pathlib.Path, start_webviz: bool = True):
         )
 
 
-def flownet_ahm(args: argparse.Namespace) -> None:
+def flownet_ahm_esmda(args: argparse.Namespace) -> None:
     """
-    Entrypoint for running flownet in AHM mode.
+    Entrypoint for running flownet in AHM mode using ES-MDA.
 
     Args:
         args: input namespace from argparse
@@ -64,7 +65,7 @@ def flownet_ahm(args: argparse.Namespace) -> None:
             )
 
     config = parse_config(args.config)
-    run_flownet_history_matching(config, args)
+    run_esmda(config, args)
 
     if not args.skip_postprocessing:
         create_webviz(args.output_folder, start_webviz=args.start_webviz)
@@ -91,6 +92,32 @@ def flownet_pred(args: argparse.Namespace) -> None:
 
     config = parse_pred_config(args.config)
     run_flownet_prediction(config, args)
+
+    if not args.skip_postprocessing:
+        create_webviz(args.output_folder, start_webviz=args.start_webviz)
+
+
+def flownet_ahm_hyperopt(args: argparse.Namespace) -> None:
+    """
+    Entrypoint for running flownet in AHM mode using hyperopt.
+
+    Args:
+        args: input namespace from argparse
+
+    Returns:
+        Nothing
+
+    """
+    if args.output_folder.exists():
+        if args.overwrite:
+            shutil.rmtree(args.output_folder)
+        else:
+            raise ValueError(
+                f"{args.output_folder} already exists. Add --overwrite or change output folder."
+            )
+
+    config = parse_config(args.config)
+    run_hyperopt(config, args)
 
     if not args.skip_postprocessing:
         create_webviz(args.output_folder, start_webviz=args.start_webviz)
@@ -130,34 +157,80 @@ def main():
         "ahm", help="Run flownet in an assisted history matching setting."
     )
 
-    parser_ahm.add_argument(
+    subparsers_ahm = parser_ahm.add_subparsers(
+        help="The options available. "
+        'Type e.g. "flownet --help" '
+        "to get help on that particular "
+        "option."
+    )
+
+    # Add assisted history matching esmda argument parser:
+    parser_esmda = subparsers_ahm.add_parser(
+        "esmda", help="Run flownet in an assisted history matching setting."
+    )
+
+    parser_esmda.add_argument(
         "config", type=pathlib.Path, help="Configuration file with AHM settings to use."
     )
-    parser_ahm.add_argument(
+    parser_esmda.add_argument(
         "output_folder", type=pathlib.Path, help="Folder to store AHM output."
     )
-    parser_ahm.add_argument(
+    parser_esmda.add_argument(
         "--overwrite",
         action="store_true",
         help="Overwrite output directory if it already exists",
     )
-    parser_ahm.add_argument(
+    parser_esmda.add_argument(
         "--skip-postprocessing",
         action="store_true",
         help="Do not run any postprocessing after ERT is finished",
     )
-    parser_ahm.add_argument(
+    parser_esmda.add_argument(
         "--start-webviz",
         action="store_true",
         help="Start webviz automatically. This flag has no effect if also --skip-postprocessing is used",
     )
-    parser_ahm.add_argument(
+    parser_esmda.add_argument(
         "--debug",
         action="store_true",
         help="Keeps all files generated by ERT for easier inspection of results during debugging",
     )
 
-    parser_ahm.set_defaults(func=flownet_ahm)
+    parser_esmda.set_defaults(func=flownet_ahm_esmda)
+
+    # Add assisted history matching hyperopt argument parser:
+    parser_hyperopt = subparsers_ahm.add_parser(
+        "esmda", help="Run flownet in an assisted history matching setting."
+    )
+
+    parser_hyperopt.add_argument(
+        "config", type=pathlib.Path, help="Configuration file with AHM settings to use."
+    )
+    parser_hyperopt.add_argument(
+        "output_folder", type=pathlib.Path, help="Folder to store AHM output."
+    )
+    parser_hyperopt.add_argument(
+        "--overwrite",
+        action="store_true",
+        help="Overwrite output directory if it already exists",
+    )
+    parser_hyperopt.add_argument(
+        "--skip-postprocessing",
+        action="store_true",
+        help="Do not run any postprocessing after ERT is finished",
+    )
+    parser_hyperopt.add_argument(
+        "--start-webviz",
+        action="store_true",
+        help="Start webviz automatically. This flag has no effect if also --skip-postprocessing is used",
+    )
+    parser_hyperopt.add_argument(
+        "--debug",
+        action="store_true",
+        help="Keeps all files generated by ERT for easier inspection of results during debugging",
+    )
+
+    parser_hyperopt.set_defaults(func=flownet_ahm_hyperopt)
 
     # Add assisted history matching argument parser:
     parser_pred = subparsers.add_parser(
