+++ src/flownet/ahm/__init__.py
@@ -4,4 +4,5 @@ from ._assisted_history_matching import (
     save_iteration_parameters,
 )
 from ._ahm_iteration_analytics import save_iteration_analytics
-from ._run_ahm import run_flownet_history_matching
+from ._run_esmda import run_esmda
+from ._run_hyperopt import run_hyperopt
